Moved TouchImageView.java to another package and added header
