+++ app/src/main/res/layout/item_fullscreen_image.xml
@@ -4,7 +4,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" >
 
-        <org.piwigo.helper.TouchImageView
+        <org.piwigo.ui.photoviewer.TouchImageView
             android:id="@+id/imgDisplay"
             android:layout_width="fill_parent"
             android:layout_height="fill_parent"
